% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataRequest.R
\name{dataRequest}
\alias{dataRequest}
\title{Get station data for specified parameter(s) and station(s)}
\usage{
dataRequest(climateParams, climateStations, sdate, edate)
}
\arguments{
\item{climateParams}{A list of one or more climate parameters (e.g. pcpn, mint, maxt, avgt, obst, snow, snwd, cdd, hdd, gdd).  See Table 3 on ACIS Web Services page: http://www.rcc-acis.org/docs_webservices.html}

\item{climateStations}{A list of one or more unique identifiers for climate stations}

\item{sdate}{Start date as a string (yyyy-mm-dd or yyyymmdd). The beginning of the desired date range. If period of record is desired, use "por".}

\item{edate}{End date as a string (yyyy-mm-dd or yyyymmdd). The end of the desired date range. If period of record is desired, use "por".}
}
\value{
A data frame containing the requested data
}
\description{
Takes a list of one or more parameters and one or more stations, requests station data, and returns it as a data frame
}

