% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationFinder.R
\name{findStation}
\alias{findStation}
\title{Find stations near a park}
\usage{
findStation(parkCodes, climateParams)
}
\arguments{
\item{parkCodes}{One or more NPS park codes as a List}

\item{climateParams}{A list of one or more climate parameters (e.g. pcpn, mint, maxt, avgt, obst, snow, snwd, cdd, hdd, gdd).  See Table 3 on ACIS Web Services page: http://www.rcc-acis.org/docs_webservices.html}
}
\value{
A data frame containing station information for stations near the specified park(s)
}
\description{
Takes one or more park codes and one or more climate parameters, determines the stations near (within ~30km) the specified park, and returns station information as a data frame with the following items: name, longitude/latitude (ll), station IDs (sids), state code, elevation (feet), and unique ID
}

