% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridRequestor.R
\name{getDailyGrids}
\alias{getDailyGrids}
\title{Get daily climate data grids for specified parameter(s) and NPS unit(s)}
\usage{
getDailyGrids(unitCode = NULL, sdate = NULL, edate = NULL,
  distance = NULL, climateParameters = NULL, filePath = NULL)
}
\arguments{
\item{distance}{(optional) Distance (in kilometers) to buffer park bounding box}

\item{sDate}{sdate (required) Format as a string (yyyy-mm-dd, yyyymmdd, yyyy-mm, yyyymm, yyyy). The beginning of the desired date range.}

\item{eDate}{edate (required) Format as a string (yyyy-mm-dd, yyyymmdd, yyyy-mm, yyyymm, yyyy). The end of the desired date range.}
}
\value{
ASCII-formatted grid file for each parameter
}
\description{
Takes one park code and a list of one or more climate parameters and requests daily climate data. Returns a grid or grids (by parameter) in ASCII format.
}

