% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{outputAscii}
\alias{outputAscii}
\title{outputAscii formats grid(s) as ASCII (*.asc) with headers and projection (*.prj)}
\usage{
outputAscii(gridResponse, fullFilePath, bbox, luSource)
}
\arguments{
\item{gridResponse}{grid (dataframe format) returned from ACIS request}

\item{bbox}{bounding box for grid request}

\item{luSource}{ACIS lookup source (as dataframe)}

\item{filePath}{full file path for ASCII output}
}
\description{
outputAscii formats grid(s) as ASCII (*.asc) with headers and projection (*.prj)
}

