% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataRequest.R
\name{dataRequest}
\alias{dataRequest}
\title{Get station data for specified parameter(s) and station(s)}
\usage{
dataRequest(dataURL, climateParams, climateStations)
}
\arguments{
\item{dataURL}{URL for ACIS data service vending station data}

\item{climateParams}{A list of one or more climate parameters (tmin, tmax, tavg, prcp)}

\item{climateStations}{A list of one or more climate stations}
}
\value{
A JSON object containing the requested data
}
\description{
Takes a list of one or more parameters and one or more stations, requests station data, and returns it as a JSON object
}

