% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationFinder.R
\name{findStation}
\alias{findStation}
\title{Find stations near a park}
\usage{
findStation(parkCodes)
}
\arguments{
\item{parkCodes}{One or more NPS park codes as a List}
}
\value{
A list of station identfiers for stations near the specified parks
}
\description{
Takes one or more park codes, determines the stations near the specified park, and returns the list of station identifiers
}

