% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationFinder.R
\name{findStation}
\alias{findStation}
\title{Find stations near a park}
\usage{
findStation(sourceURL, parkCode)
}
\arguments{
\item{sourceURL}{sourceURL for ACIS data services}

\item{parkCode}{NPS park code}
}
\value{
A list of stations near the specified tatpark
}
\description{
Takes a sourceURL for ACIS data services, determines the stations near the specified park, and returns the list of station identifiers
}

