% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridRequestor.R
\name{getDailyGrids}
\alias{getDailyGrids}
\title{Get daily climate data grids for specified parameter(s) and NPS unit(s)}
\usage{
getDailyGrids(unitCode = NULL, sdate = NULL, edate = NULL,
  distance = NULL, climateParameters = NULL, filePath = NULL)
}
\arguments{
\item{unitCode}{(optional) One unit code as a string. If not specified, defaults to CONUS geographic extent.}

\item{distance}{(optional) Distance (in kilometers) to buffer park bounding box}

\item{climateParameters}{(optional) A list of one or more climate parameters (e.g. pcpn, mint, maxt, avgt).  If not specified, defaults to all parameters. See the ACIS Web Services page: http://www.rcc-acis.org/docs_webservices.html}

\item{filePath}{filePath (optional) Folder path for output ASCII grid(s). If specified, grid(s) are saved to the folder. Otherwise, grid(s) are saved to the current working directory.}

\item{sDate}{sdate (required) Format as a string (yyyy-mm-dd, yyyymmdd, yyyy-mm, yyyymm, yyyy). The beginning of the desired date range.}

\item{eDate}{edate (required) Format as a string (yyyy-mm-dd, yyyymmdd, yyyy-mm, yyyymm, yyyy). The end of the desired date range.}
}
\value{
ASCII-formatted grid file for each parameter
}
\description{
Takes one park code and a list of one or more climate parameters and requests daily climate data. Returns a grid or grids (by parameter) in ASCII format.
}
\examples{
\dontrun{
Two daily grids for PRWI for one date: returns one grid for each parameter for each date - 4 grids total
getDailyGrids(unitCode = list("PRWI"), sdate = "20160615", edate = "20160616", climateParameters = list("mint", "maxt"), filePath="d:\\\\temp\\\\trash")

Two daily grids for GRSM for one date: print output to console
getDailyGrids(unitCode = list("GRSM"), sdate = "20160615", edate = "20160616", climateParameters = list("mint", "maxt"))
}
}

