% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataRequest.R
\name{getDailyWxObservations}
\alias{getDailyWxObservations}
\title{Get station data for specified parameter(s) and station(s)}
\usage{
getDailyWxObservations(climateParams, climateStations, sdate = "por",
  edate = "por", filePathAndName = NULL)
}
\arguments{
\item{climateParams}{A list of one or more climate parameters (e.g. pcpn, mint, maxt, avgt, obst, snow, snwd, cdd, hdd, gdd).  See Table 3 on ACIS Web Services page: http://www.rcc-acis.org/docs_webservices.html}

\item{climateStations}{A list of one or more unique identifiers for climate stations}

\item{sdate}{(optional) Default is period of record ("por"). If specific start date is desired, format as a string (yyyy-mm-dd or yyyymmdd). The beginning of the desired date range.}

\item{edate}{(optional) Default is period of record ("por"). IF specific end date is desired, format as a string (yyyy-mm-dd or yyyymmdd). The end of the desired date range.}

\item{filePathAndName}{(optional) File path and name including extension for output CSV file}
}
\value{
A data frame containing the requested data
}
\description{
Takes a list of one or more parameters and one or more unique station IDs, requests station data, and returns it as a data frame
}

