% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{outputAscii}
\alias{outputAscii}
\title{outputAscii formats grid(s) as ASCII (*.asc) with headers and projection (*.prj)}
\usage{
outputAscii(gridResponse)
}
\arguments{
\item{gridResponse}{grid (JSON format) returned from ACIS request}
}
\description{
outputAscii formats grid(s) as ASCII (*.asc) with headers and projection (*.prj)
}

